RealEstateChatBot/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
Django==5.2.8
django-cors-headers==4.9.0
et_xmlfile==2.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
groq==0.35.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jiter==0.12.0
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
